backend/requirements.txt
# Biesse Technical Documentation Chat Assistant - Backend Dependencies

# AI/ML Stack
google-generativeai==0.3.2
chromadb==0.4.22

# PDF Processing
pymupdf==1.23.8

# Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database (for future use)
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
loguru==0.7.2

# Development
pytest==7.4.0
pytest-asyncio==0.23.0
httpx==0.26.0
